move rest of the docker service
